backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
drf-yasg
pillow
mysqlclient
django-debug-toolbar
channels
channels-redis